Move agent special permissions tests to permissions_test.py
