+++ adaptive_hockey_federation/tests/url_test.py
@@ -1,4 +1,3 @@
-from http import HTTPStatus
 from typing import Any
 
 import pytest
@@ -161,86 +160,3 @@ class TestUrls(TestCase):
         """Тест - удаление пользователя."""
         delete_result = self.delete_user(self.user.id)
         self.assertTrue(delete_result, "Ошибка при удалении пользователя")
-
-    def test_agent_has_access(self):
-        """Доступ представителя к своей команде, ее игрокам и т.д."""
-        # TODO: Добавить урлы тренеров и пушер-тьюторов по аналогии,
-        #  когда функционал ограничения доступа к этим сущностям будет
-        #  разработан.
-        urls_agent_has_access_to = {
-            f"/teams/{self.team_2.id}/edit/": (
-                "страница /teams/<team_id>/edit/ редактирования общих "
-                "сведений этой команды (ожидается ответ со статусом 200)"
-            ),
-            f"/players/{self.player_2.id}/": (
-                "страница просмотра подробных сведений об игроке из этой "
-                "команды (страница /players/<player_id>/ должна вернуть "
-                "ответ со статусом 200)"
-            ),
-            f"/players/{self.player_2.id}/edit/": (
-                "страница редактирования игрока из этой команды (страница "
-                "/players/<player_id>/edit/ должна вернуть ответ со статусом "
-                "200)"
-            ),
-            f"/players/create/?team={self.team_2.id}": (
-                "страница создания игрока из этой команды (страница "
-                "/players/create/?team=<team_id> должна вернуть ответ со "
-                "статусом 200)"
-            ),
-        }
-        self.client.force_login(self.user_agent)
-        for url, message in urls_agent_has_access_to.items():
-            with self.subTest(msg=message, url=url):
-                response = self.client.get(url)
-                self.assertEqual(
-                    response.status_code,
-                    HTTPStatus.OK,
-                    msg=(
-                        "Представителю команды должна "
-                        "быть доступна " + message
-                    ),
-                )
-
-    def test_agent_has_no_access(self):
-        """Запрет доступа представителя к чужой команде, игрокам и т.д."""
-        # TODO: Добавить урлы тренеров и пушер-тьюторов по аналогии,
-        #  когда функционал ограничения доступа к этим сущностям будет
-        #  разработан.
-        urls_agent_has_no_access_to = {
-            f"/teams/{self.team.id}/edit/": (
-                "страница /teams/<team_id>/edit/ редактирования общих "
-                "сведений ЧУЖОЙ команды (ожидается ответ со статусом 403)"
-            ),
-            f"/players/{self.player.id}/": (
-                "страница просмотра подробных сведений об игроке ЧУЖОЙ "
-                "команды (страница /players/<player_id>/ должна вернуть "
-                "ответ со статусом 403)"
-            ),
-            f"/players/{self.player.id}/edit/": (
-                "страница редактирования игрока ЧУЖОЙ команды (страница "
-                "/players/<player_id>/edit должна вернуть ответ со статусом "
-                "403)"
-            ),
-            f"/players/create/?team={self.team.id}": (
-                "страница создания игрока с привязкой к ЧУЖОЙ команде "
-                "(страница /players/create/?team=<team_id> должна вернуть "
-                "ответ со статусом 403)"
-            ),
-            "/players/create/": (
-                "страница создания игрока без привязки к какой-либо команде "
-                "(страница /players/create/ должна вернуть ответ со "
-                "статусом 403)"
-            ),
-        }
-        self.client.force_login(self.user_agent)
-        for url, message in urls_agent_has_no_access_to.items():
-            with self.subTest(msg=message, url=url):
-                response = self.client.get(url)
-                self.assertEqual(
-                    response.status_code,
-                    HTTPStatus.FORBIDDEN,
-                    msg=(
-                        "Представителю команды НЕ должна "
-                        "быть доступна " + message
-                    ),
-                )
